backend/requirements.txt
sentence-transformers
faiss-cpu
transformers==4.39.3
torch
accelerate
gradio
fastapi
uvicorn[standard]
pydantic
python-multipart